feat: Implement Price Management Module
- Added PriceRepository for CRUD operations on Price entity.
- Created PriceScheduledService for automatic price status updates.
- Developed PriceService interface and its implementation for managing price-related business logic.
- Introduced methods for creating, updating, retrieving, and deleting prices.
- Implemented validation for price creation and updates, including status transitions.
- Added functionality to handle price details and ensure no conflicts with current prices.
- Scheduled tasks to activate upcoming prices and expire current prices based on time conditions.

+++ src/main/java/iuh/fit/supermarket/entity/BannerVariantId.java
@@ -14,14 +14,14 @@ import java.io.Serializable;
 @NoArgsConstructor
 @AllArgsConstructor
 public class BannerVariantId implements Serializable {
-    
+
     /**
      * ID banner
      */
     private Integer banner;
-    
+
     /**
      * ID biến thể sản phẩm
      */
-    private Integer variant;
+    private Long variant;
 }
